+++ src/app/(myReview)/searchReviews/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import ReviewText from "@/components/addReview/ReviewText";
-import ReviewHeader from "@/components/mainPage/ReviewHeader";
-import ReviewInfo from "@/components/mainPage/ReviewInfo";
+import ReviewText from "@/components/ui/review/ReviewText";
+import ReviewHeader from "@/components/ui/review/ReviewHeader";
+import ReviewInfo from "@/components/ui/review/ReviewInfo";
 import useSearchReviews from "@/hooks/useSearchReviews";
 import { Review } from "@/types/addReview";
 import { useSearchParams } from "next/navigation";
