+++ src/components/header/UserMenu.tsx
@@ -2,10 +2,10 @@ import { signOut } from "firebase/auth";
 import auth from "@/lib/firebase/firebase";
 import Link from "next/link";
 import { User } from "@/lib/store/user";
-import Dropdown from "../ui/dropdown/Dropdown";
-import DropdownItem from "../ui/dropdown/DropdownItem";
-import DropdownList from "../ui/dropdown/DropdownList";
-import DropdownToggle from "../ui/dropdown/DropdownToggle";
+import Dropdown from "../ui/dropdown-ui/Dropdown";
+import DropdownItem from "../ui/dropdown-ui/DropdownItem";
+import DropdownList from "../ui/dropdown-ui/DropdownList";
+import DropdownToggle from "../ui/dropdown-ui/DropdownToggle";
 
 export default function UserMenu({ user }: { user: User }) {
   const handleLogout = async () => {
