+++ src/components/ui/VoteAverage.tsx
