+++ src/components/ui/movie/MovieCarousel.tsx
